Move winston-based logger to logger.node
